bugfix: otp-433: Write file after cell ranger execution
At the moment, it is try to write the file to an directory not existing
at this timepoint. Therefore writing the file is moved after the cell
ranger execution.

Creating the directory before is not possible, because cell ranger assumes
it has created it, but the structure doesn't match. Therefore the
command fail.