minor: Increase logging during project creation
The creation of the analysis directory after the project creation
sometimes fails in production. When executing the commands manually it
seems fine but the error message only shows stdout and not stderr.

Now it logs the call in more details and in the error message you can
now see stdout, stderr and the exit code.