new: Introduce GUI_ANNOTATION for easy location specific annotations
The ProcessingOption GUI_ANNOTATION and its subtypes allow for
configurable and easy to introduce location specific GUI annotations.

New annotations can be simply integrated by adding a value to the enum
GuiAnnotation and placing the TagLib component with the respective type
at the expected location in the GSP.

The tag simply expects the subtype to get its content from and the type
of annotation to wrap it in. (see: otp:annotation)

The content is rendered as HTML.