otp-429: handle both TSV and CSV in metadata imports
We introduce basic auto-detection of which delimiter to use, based on a
best-effort guess when looking at the header.
If we find something actively annoying in the header, give up with an
Exception.
If we just don't know, fall back to the previous default: TAB.

- Refactoring: improve Spreadsheet docu, and release resources when done.
- Refactoring: extract Delimiter to own file, it was getting too big to
  be maintainable.
- Provide an enum value to say we want to auto-detect delimiters
  and make sure the rest of OTP doesn't trip over this addition. (see
  note 1 below)
- switch metadata import to autodetect delimiters.

Note 1) The "auto_detect" value changes the enum from a simple "data holder"
to something with additional semantics/meaning (but only for the auto-detect
value). We must isolate the parts of OTP that don't know how to handle this
(yet) from the new value, and include safeguards that blow up loudly if we
miss something.
For this reason, some 'IllegalArgumentException' are thrown if the
wrong properties are accessed on the wrong Enum-option.