Refactor: Document implicit expectation for testing group property.
When setting up OTP in a new environment, I naively set the
testing-group property to the laziest value I could get away with:
my current primary group, so I wouldn't have to make new ones.

It took me more than an hour of debugging and stack-tracing to
figure out why tests were failing locally, but not on CI or my
work PC.

Lets explicitly document this "common knowledge"/requirement on
the testing properties with an assertion and some clear messages.