otp-436: Extend ClusterJob and Workflow for new WF-System
Small extensions for ClusterJob and Workflow in preparation for the new
workflow system.

ClusterJobs now belong to either the old or new system, defaulting to
the old one for now.
Workflows got a priority to be used for scheduling, which is set to 0
for now but will be set to its actual values at a later point in time.