otp-541: switch user deactivation check to LDAP UserAccountControl
Implemented an enum covering all known values of the LDAP value
UserAccountControl. The check for deactivated users is now based on the
value ACCOUNTDISABLE.

A table with all values and their state for a user is shown on the page
"userAdministration/show".
On the same page there is also a warning now, when the user can not be
found in LDAP.

As the deactivation criteria has been determined the cronjob
ScheduleUsersForDeactivationJob has been reactivated.

The job now now also checks all users with a pending deactivation date,
regardless of project membership. This is to reset the dates for people
which were scheduled, but have since been removed from all projects.