otp-584: Remove manipulated authentication in ProcessesController
The asynchronously used function processData in the ProcessesController
now uses tasks for threading and resets the authentication in each
thread with a 'finally' statement.
This caused issues as the authentication permanently stays in the the
affected thread of the pool. In turn, when this thread is used again,
the authentication object is still set according to the already finished
function.

Now the authentication object will always be removed, keeping the thread
clean.