chg: otp-569: Define MetaDataColumn alias SAMPLE_NAME for SAMPLE_ID
As the preferred term in the development team and also for affected
stakeholders is SAMPLE_NAME we decided to use SAMPLE_NAME as the true
value and SAMPLE_ID as its alias.

This commit contains the rename of the enum value, the alias and the
migration. For the renames in the rest of the codebase there are
separate commits to keep it at least a bit organized.