new: otp-570: Allow sending a single QC issue mail for multiple bams
The function createResultsAreBlockedMessage is now overloaded and also
provides a signature with a bam list parameter.

This function uses its own template as it is not really possible to
reuse the single-bam one.

The text is rephrased to match the multi-bam nature of the text.

Bams are grouped by their SeqType and for each group a SeqType specific
link to the alignment quality overview is provided.

Affected bams are then shown in an indented list. For each bam we
include sample specific links to the alignment quality overview and
result directory on the filesystem.

All components of the mail have been exposed for further use.