new: otp-570: Allow manual notification on import ticket basis
The metadata import details view has been extended with a selection form
for processing steps and the QC traffic light notification. Upon
submitting a notification digest for all selected notifications is
generated and displayed on the new notification preview view.

The digest reuses parts of the typical notification but sends out a
single combined mail instead of many.