new: otp-588: migrate execute pipeline job: Add a log service to create new log entries for the new workflow system.
The logs are created in a seperate tranaction, so they are persistent also in the case, that the origin transaction are rolled back.

For create a new transaction the helper TransactionUtils is added, since that simplefies mocking in tests.

Also the existing test using WorkflowMessageLog is updated to use the logservice