new: otp-588: migrate execute pipeline job: change connection directory between workflow step and logs
Currently, the workflow step contains a list of logs. But than it is not possible to add persistence logs
and rollback changes in workflow step. Therefore, the direction of connection is changed:
- A WorkflowStep is added to the WorkflowStep
- The logs property is removed from WorkflowStep
- A method to get the logs of a WorkflowStep is added to the WorkflowStep

This allow to add logs in a seperate transaction, so a rollback of the step do not remove the logs.