new: otp-338: Provide a way to block actions for switched users
Provides the tag otpsecurity:noSwitchedUser and associated function in
the new SecurityController to make the function throw an exception.

To block something we need to secure the html and groovy code.

For the gsp, wrap the associated submit buttons with the new tag
noSwitchedUser and include the javascript `taglib/NoSwitchedUser.js`.

In the Groovy code, call `SecurityService.ensureNotSwitchedUser`
wherever you want to assert that the user is not switched. This function
will throw a `SwitchedUserDeniedException` in case the user is switched.

These exceptions are handled in the UrlMapping to redirect to a
dedicated view.