fix: otp-700: Show empty Samples on the lane overview
Before recent changes regarding the Sample Identifier, "empty" Samples
were listed in the lane overview, just without any data in the later
columns.

Because this was pretty much an unintended side effect, it was removed
as part of the sample identifier removal.

This commit contains a dirty fix to show them again, by simply adding
all Samples of the Project to the list.