chg: otp-460: Changes to user management in project request
The restriction of one PI (approver) per request has been liften. In its
place there is now a multi approver system, which waits for a list of
approvers to either approve or deny.

Who is expected to approve is based on the authority roles, so currently
only PIs will be expected to approve or deny.

Furthermore the entire system to add users has been changed to allow
providing multiple roles and also some permissions per user.

To escape from eternally unapproved requests, a requester now has the
possibility to close a request.

A request can be edited by the requester. Upon an edit, the approvals
will be reset and the approvers will be notified that there have been
changes and their approval is required again.

Release Notes:
The changes in the structure of the project request will make all
existing requests invalid. Make sure you deleted all existing requests.