bug: otp-756: add check of cluster jobs in validation jobs
Since there are exist different pipeline types (otp, roddy, wes), different implementation to
check, if the pipeline run through are needed. Therefore a subclass for each type was introduced, and for
otp pipelines already implemented. The other both subclasses are added only as skeletion.

The class AbstractValidationJob was adapted to
- introduced the new callback method, implemented in the three new subclasses
- include the call in the check procedure
- improve the code a little