chg: otp-878: Set chmod to some parent directory of roody workflows
In one location, the permissions of the result directoy '*_result' are not set correctly.
Therefore the code to create the directory are adapted to set the permission to all of the levels.
Since the code is shared between all roddy workflows, also the roddy alignments (PANCAN, WGBS PANCAN, and rna alignment)
is effected, but that shouldn't cause problems, since the directories there should have that permission also.