chg: otp-596: prework: Introduce check and clean methods to FileSystemService and RemoteShellHelper
For workflow tests it is necessary to remove old remote accesses, since in the next test it is connected to other objects.
In the old system its done via direct access to the methods, which is not nice.

Therefore I added now new methods therefore. Nevertheless I don't adapted to old workflow system, since that will be deleted later.