bugfix: otp-1360: Fix problems with realm
After changing otp paths to use always remote path, the test access the realm before it was corrected for remote access.

Therefore, the test overwrite the createRealm methods of the traits to ensure the correct values and also to ensure only one is created.

There were also a problem in the AbstractRunYapsaWorkflowTests, which was fixed.