chg: otp-1594: new level for fastqc
Release note:
Since a new fastqc level is introduce, a migration is needed. The database is migrated automatically,
but the file system needs to be migrated manually. The groovy script otp-1594-new-fastqc-level generates
a bash script per project and one to call them all. The generation can take some minutes.