bug: otp-2163: fix bug of double started job
the restart job has create the restarted job, but also changed the run state to pending, which is only for not yet started workflows with some run jobs.
So there were a race condition between the queued job and the run.
If the job executes first, it change the state of run to running, so it won't be fetched by start run.
But if the run was fetched first, it creates another job of the workflow, so now two are in queue to start.
Therefore, the restart job do not change the run state anymore, that is then done by the start job.

Also, some additianal asserts were added.