bug: otp-2617: Fix cellranger plot directory
During the fix of to long runids, the name of runids were changed.
Since the path is constructed on the fly, the work path is now constructed constructed wrongly.
The files are also linked to a final place, which has not changed, but
the gui use the work directory. Since the gui shows the plots also before
a final run is selected, it needs now know both cases.