bug: otp-2817: Handle trigger cases, were not both alignments part of the given list
Do the bug, only the categories of the given list were fetched, but if the corresponding bam file was not part of it and
also no other input has that sample type, the category was not fetched and therefor unknown. Therefore otp asume,
there is no pair and doesn't create sample pairs.