bug: otp-2811: handle skip workflows for notification
currently, the artefacts of skipped notification stays, so that the notification system thinks, there are more workflows.
To solve it, the code is adapted to mark them as woithdrawn, so that the notification system ignore them and don't wait therefore anymore.

Since there was no common interface to do so, there were some coded needed to get the correct service for the possible domain classes to find the correct service.